HW2/requirements.txt
accelerate>=0.30.0
transformers>=4.44.0

openmim>=0.3.9
mmengine>=0.10.4
mmdet>=3.3.0
mmcv>=2.0.0rc4,<2.2.0
terminaltables>=3.1.10

numpy>=1.26.0
pandas>=2.0.0
matplotlib>=3.8.0
Pillow>=10.0.0
tqdm>=4.65.0
pycocotools>=2.0.7
scikit-learn>=1.3.0
